重构

+++ betterme-framework/code/betterme-framework-model/src/main/java/com/lgfei/betterme/framework/model/vo/PageResultVO.java
@@ -4,23 +4,28 @@ import java.util.List;
 
 public class PageResultVO<T>
 {
-	private List<T> rows;
-	private Integer total;
+    private List<T> rows;
+    
+    private Long total;
+    
     public List<T> getRows()
     {
         return rows;
     }
+    
     public void setRows(List<T> rows)
     {
         this.rows = rows;
     }
-    public Integer getTotal()
+    
+    public Long getTotal()
     {
         return total;
     }
-    public void setTotal(Integer total)
+    
+    public void setTotal(Long total)
     {
         this.total = total;
     }
-	
+    
 }
